requirements.txt
tensorflow
gradio==3.50
Pillow
